reorganize weather/
