+++ src/components/weather/ObservationPanel.tsx
